# apps/web-client/src/types/api/entity-schemas.types.ts
import type { components } from './schema'

export type FieldType = components['schemas']['FieldType']
export type EntityFieldItem = components['schemas']['EntityFieldItem']
export type EntitySchemaListItem = components['schemas']['EntitySchemaListItem']
export type CreateEntitySchemaRequest = components['schemas']['CreateEntitySchemaRequest']
export type CreateEntitySchemaResponse = components['schemas']['CreateEntitySchemaResponse']
export type GetEntitySchemaListResponse = components['schemas']['GetEntitySchemaListResponse']
export type GetEntitySchemaResponse = components['schemas']['GetEntitySchemaResponse']
export type UpdateEntitySchemaRequest = components['schemas']['UpdateEntitySchemaRequest']