# apps/web-client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MazyPlatform.Service.Gateway",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/bots": {
      "post": {
        "tags": [
          "BotService"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBotRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBotResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "BotService"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBotsByUserIdResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bots/standalone": {
      "post": {
        "tags": [
          "BotService"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBotWithoutProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBotResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bots/{botInstanceId}": {
      "get": {
        "tags": [
          "BotService"
        ],
        "parameters": [
          {
            "name": "botInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBotResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "BotService"
        ],
        "parameters": [
          {
            "name": "botInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empty"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/bots": {
      "get": {
        "tags": [
          "BotService"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBotsByProjectResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bots/{botInstanceId}/bind": {
      "post": {
        "tags": [
          "BotService"
        ],
        "parameters": [
          {
            "name": "botInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindBotToProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empty"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bots/{botInstanceId}/unbind": {
      "post": {
        "tags": [
          "BotService"
        ],
        "parameters": [
          {
            "name": "botInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnbindBotFromProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empty"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bots/{botInstanceId}/activate": {
      "post": {
        "tags": [
          "BotService"
        ],
        "parameters": [
          {
            "name": "botInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateBotRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empty"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bots/{botInstanceId}/deactivate": {
      "post": {
        "tags": [
          "BotService"
        ],
        "parameters": [
          {
            "name": "botInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeactivateBotRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empty"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bots/{botInstanceId}/token": {
      "put": {
        "tags": [
          "BotService"
        ],
        "parameters": [
          {
            "name": "botInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBotTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empty"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/schemas": {
      "post": {
        "tags": [
          "EntitySchemaService"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntitySchemaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateEntitySchemaResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "EntitySchemaService"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEntitySchemaListResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schemas/{schemaId}": {
      "delete": {
        "tags": [
          "EntitySchemaService"
        ],
        "parameters": [
          {
            "name": "schemaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empty"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "EntitySchemaService"
        ],
        "parameters": [
          {
            "name": "schemaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEntitySchemaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empty"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "EntitySchemaService"
        ],
        "parameters": [
          {
            "name": "schemaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEntitySchemaResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivateBotRequest": {
        "type": "object",
        "properties": {
          "botInstanceId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Any": {
        "required": [
          "@type"
        ],
        "type": "object",
        "properties": {
          "@type": {
            "type": "string"
          }
        },
        "additionalProperties": {
          "$ref": "#/components/schemas/Value"
        }
      },
      "BindBotToProjectRequest": {
        "type": "object",
        "properties": {
          "botInstanceId": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "scenarioVersion": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "BotListItem": {
        "type": "object",
        "properties": {
          "botInstanceId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "platformType": {
            "$ref": "#/components/schemas/BotPlatformType"
          },
          "maskedAccessToken": {
            "type": "string"
          },
          "communityId": {
            "type": "string"
          },
          "scenarioVersion": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/BotStatus"
          },
          "projectId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "BotPlatformType": {
        "enum": [
          "Bot_PLATFORM_TYPE_UNSPECIFIED",
          "Bot_PLATFORM_TYPE_VK",
          "Bot_PLATFORM_TYPE_TELEGRAM"
        ],
        "type": "string"
      },
      "BotStatus": {
        "enum": [
          "BOT_STATUS_UNSPECIFIED",
          "BOT_STATUS_INACTIVE",
          "BOT_STATUS_ACTIVE"
        ],
        "type": "string"
      },
...
      "CreateBotRequest": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "platformType": {
            "$ref": "#/components/schemas/BotPlatformType"
          },
          "accessToken": {
            "type": "string"
          },
          "communityId": {
            "type": "string"
          },
          "scenarioVersion": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CreateBotResponse": {
        "type": "object",
        "properties": {
          "botInstanceId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CreateBotWithoutProjectRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "platformType": {
            "$ref": "#/components/schemas/BotPlatformType"
          },
          "accessToken": {
            "type": "string"
          },
          "communityId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CreateEntitySchemaRequest": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CreateEntitySchemaResponse": {
        "type": "object",
        "properties": {
          "schemaId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "DeactivateBotRequest": {
        "type": "object",
        "properties": {
          "botInstanceId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Empty": {
        "type": "object",
        "additionalProperties": false
      },
      "EntityFieldItem": {
        "type": "object",
        "properties": {
          "fieldId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "fieldType": {
            "$ref": "#/components/schemas/FieldType"
          },
          "isRequired": {
            "type": "boolean"
          },
          "defaultValue": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "EntitySchemaListItem": {
        "type": "object",
        "properties": {
          "schemaId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "FieldType": {
        "enum": [
          "FIELD_TYPE_UNSPECIFIED",
          "FIELD_TYPE_STRING",
          "FIELD_TYPE_NUMBER",
          "FIELD_TYPE_BOOLEAN",
          "FIELD_TYPE_DATE_TIME",
          "FIELD_TYPE_REFERENCE",
          "FIELD_TYPE_ENUM"
        ],
        "type": "string"
      },
      "GetBotResponse": {
        "type": "object",
        "properties": {
          "botInstanceId": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "platformType": {
            "$ref": "#/components/schemas/BotPlatformType"
          },
          "maskedAccessToken": {
            "type": "string"
          },
          "communityId": {
            "type": "string"
          },
          "scenarioVersion": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/BotStatus"
          }
        },
        "additionalProperties": false
      },
      "GetBotsByProjectResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BotListItem"
            }
          }
        },
        "additionalProperties": false
      },
      "GetBotsByUserIdResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BotListItem"
            }
          }
        },
        "additionalProperties": false
      },
      "GetEntitySchemaListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntitySchemaListItem"
            }
          }
        },
        "additionalProperties": false
      },
      "GetEntitySchemaResponse": {
        "type": "object",
        "properties": {
          "schemaId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityFieldItem"
            }
          }
        },
        "additionalProperties": false
      },
...
      "KindOneofCase": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ListValue": {
        "type": "object",
        "properties": {
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Value"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "NullValue": {
        "enum": [
          0
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Status": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Any"
            }
          }
        },
        "additionalProperties": false
      },
      "Struct": {
        "type": "object",
        "properties": {
          "fields": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Value"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "UnbindBotFromProjectRequest": {
        "type": "object",
        "properties": {
          "botInstanceId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UpdateBotTokenRequest": {
        "type": "object",
        "properties": {
          "botInstanceId": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          },
          "communityId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UpdateEntitySchemaRequest": {
        "type": "object",
        "properties": {
          "schemaId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "fieldType": {
            "$ref": "#/components/schemas/FieldType"
          },
          "isRequired": {
            "type": "boolean"
          },
          "defaultValue": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Value": {
        "type": "object",
        "properties": {
          "nullValue": {
            "$ref": "#/components/schemas/NullValue"
          },
          "hasNullValue": {
            "type": "boolean",
            "readOnly": true
          },
          "numberValue": {
            "type": "number",
            "format": "double"
          },
          "hasNumberValue": {
            "type": "boolean",
            "readOnly": true
          },
          "stringValue": {
            "type": "string",
            "nullable": true
          },
          "hasStringValue": {
            "type": "boolean",
            "readOnly": true
          },
          "boolValue": {
            "type": "boolean"
          },
          "hasBoolValue": {
            "type": "boolean",
            "readOnly": true
          },
          "structValue": {
            "$ref": "#/components/schemas/Struct"
          },
          "listValue": {
            "$ref": "#/components/schemas/ListValue"
          },
          "kindCase": {
            "$ref": "#/components/schemas/KindOneofCase"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}